Semantic improvement in Redux
