+++ src/store/actions/__tests__/speakers.test.js
