+++ src/store/reducers/__tests__/speakers.test.js
